This patch is about bug #6602. It turns most of the QLabels in kinfocenter selectable with mouse by setting setTextInteractionFlags(Qt::TextSelectableByMouse).
This is probably no completely satisfying solution, but I
think it is an improvement. While it is not possible to
select several QLabels at once, one QLabel at a time can be
marked and copied, which is useful to easily extract e. g.
the exact amount of physical memory in bytes, the MAC address
of an NIC or the name of a hard drive or CPU.

Also, some unused or duplicate includes of QLabel are dropped.

Thanks to Martin Walch for patches.

BUG: 6602