[kinfocenter] Make GLX and EGL not mutual exclusive
GLX and EGL can both be used at the same time. What cannot be used
is OpenGL and OpenGLES at the same time. In the complete detection
code there is only one GLES specific include and one function call.
Everything else is just EGL specific. Thus it is nice to get the
information about GLX and EGL at the same time.

The compile time switches are reworked to have a KCM_HAVE_GLX which
is defined to 1 in case that we build for OpenGL (GLX is no-no if
we build for OpenGLES) and a KCM_HAVE_EGL if EGL library is found.
This can be used in both cases.

All the code which used to be ifdefed for OpenGL specific becomes
ifdefed for GLX specific and all the code specific for OpenGLES
becomes EGL specific with the exception of the eglBindAPI which
still needs to be in the GLES ifdef.

The detection code is split into two methods: one for GLX and one
for EGL and both are run depending on which defines are set.

As a slight unrelated change the name "OpenGL/ES" is split via
ifdefs in "OpenGL" and "OpenGL ES". The cases are clearly different
and that should be reflected in the UI and not mixed together.

REVIEW: 115934