Remove querying the indirect rendering context
Mesa fails to make an indirect rendering context current and freezes
in the next X11 call.

Removing indirect rendering support seems like a workaround for the
problem, but realisticly it's no longer interesting what indirect
rendering provides if direct rendering is available. If direct
rendering is not available, it will pick indirect instead.

BUG: 344971
FIXED-IN: 5.4.0
REVIEW: 123995