Pass transient parent window to KToolTipWindow
Summary:
This fixes tooltips on Wayland. We need the relationship to the
transient parent window - tooltips are positioned relative to their
parent window. Also on X11 this is an important information which was
completely lacking.

Before the change the tooltips got randomly positioned by the Wayland
compositor as it thought that it is a new toplevel window. With the
change the tooltip gets shown where it is supposed to be shown.

Test Plan: Run kinfocenter with --platform wayland and xcb

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D3113