Allow to select module with keyboard
Only the "clicked" signal was being tracked by the side panel. As a result
if the user wanted to select an element with the keyboard it would be ignored.

"activated" signal cannot be used exclusively as in that case if the user has
configured the system to require double click to open files "activated" will
not be called when a single click is done.

For that reason, both "clicked" and "activated" are tracked and
KInfoCenter::setKcm has been modified to not reload the module if it's the
same module that is already loaded.