Ensure Registry gets destroyed before the Wayland connection
Summary:
Registry was parented to WaylandModule, but the ConnectionThread got
destroyed in the dtor, thus before the Registry resulting in a crash
when the Registry got destroyed.

This change addresses this problem by tracking the Registry and
destroying it before destroying the Connection.

BUG: 387479
FIXED-IN: 5.12.1

Test Plan:
Clicked Wayland section, then quit application, verified
exit code

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D10391