Move "update" button from the left side to the right side in NIC module
Summary:
The rationale of that change is that at some other place in the KInfoCenter, the buttons are aligned to the right (for example in the IEEE 1394 device info, see the "Generate 1394 Bus Reset" button), so I think it's better if everything is aligned the same way.

Before:
{F6186087}

After:
{F6186086}

Test Plan: Run the updated KInfoCenter. Go in the Network Information / Network Interfaces module and check that the update button is in the lower right corner instead of the left corner. The rest of the functionality is unchanged.

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: abetts, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14737