Add "Copy Info" button to the About System KCM
Summary:
Adds a new button with a "Copy to Clipboard" action to the "About" System module. The action copies the visible information to the clipboard. It can be used for quickly retrieving information for a bug report.

This is an example of how the copied text looks like:

Distro: openSUSE Tumbleweed 20180404
KDE Frameworks Version: 5.46.0
Qt Version: 5.10.0
Kernel Version: 4.15.13-2-default
OS Type: 64-bit
Processors: 4 × Intel® Core™ i3 CPU M 370 @ 2.40GHz
Memory: 7,7 GiB of RAM

(Note that in this example 'KDE Plasma Version' is missing here because it is hidden in the development version)

Screenshot:

{F5691627}

BUG: 366266

Also note, that the layout in Modules/about-distro/src/Module.ui had to be reworked in order to properly integrate the new button (see screenshots below). As a nice side effect is that the behavior of horizontal resizing is now better than before. The content is now always centered instead of shifted to the right.

Reviewers: ngraham, dhaumann, rkflx

Reviewed By: ngraham, dhaumann

Subscribers: rkflx, dhaumann, ltoscano, sebas, elvisangelaccio, cfeck, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7087