only enable the module help button if the module actually has help
Summary:
the module help button specifically launches the help for DocPath set in
the desktop file of the KCM, previously the code however only checked if
the KCM pragmatically disabled the help button. while supposedly its good
form for a module to explicitly disable the help button when it has no help
but since that depends on humans to remember doing that it's a bit
unreliable.

so, the module help button now only gets enabled when there actually is
module help available AND the button wasn't explicitly disabled.
this prevents the help button from getting enabled but then launching
the generic khelpcenter since no actual DocPath was set in the kcm
desktop metadata

Test Plan:
- about (has no doc and disables button) -> disabled button
- memory (has doc) -> enabled button
- energy (has no doc but doesn't explicitly disable button) -> disabled button

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19188