connect clicked signal of view (again)
Summary:
activated, when the system is in double click mode is only triggered
on double click. but even with in double click mode we want side panels to
trigger on single clicks. so we need both clicked and activated.

double-activation continues to be properly guarded in clickedSlot, so
this change looks to be save vis-a-vis state changes in the UI

BUG: 405373
FIXED-IN: 5.15.3
CHANGELOG: single-clicks correctly activate modules again when the system is in double-click mode

Test Plan:
single click system:
- clicking enables modules
- keyboard navigation enables modules
- double clicking has no unintended side effects

double-click system:
- same

Reviewers: mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D19703