[Energy] Put battery capacity info in battery section and call it "Capacity degradation"
Summary:
Alternative to D22882.

Our use of "capacity" to refer to how much of a battery's original capacity remains is
confusing. This is made less confusing by inverting the number and referring to it as
"degradation". "Capacity degradation: 21%" communicates that the battery has lost
21% of its original capacity far better than "Capacity: 79%" does, because the latter
is easily confused with the battery's current charge level.

To further reduce the confusion, the battery's degradation level is relocated to end
of the {nav Battery} section.

Test Plan: {F7264972, size=full}

Reviewers: meven, broulik, #vdg, #plasma, filipf

Reviewed By: #vdg, filipf

Subscribers: aspotashev, filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23270