[energy kcm] Display Vendor and model
Summary:
The code is not great, I would be open to have a suggestion to improve it.
I tried to pass Solid::Device directly to the qml side and using QHash to store vendor and product but they are not QObject so it can't work.

Test Plan:
After:
{F7272955}

Reviewers: broulik, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23152