[Energy KCM] Port away from WorkspaceComponents.BatteryIcon and improve presentation
Summary:
Using this Plasma component in an app causes the icons to disappear when using a dark
Plasma theme and a light color scheme, or vice versa. This patch fixes that by porting
away from the Plasma component and instead using a Kirigami icon to show the icons.

The patch also spruces up the visual presentation a bit.

BUG: 411051
FIXED-IN: 5.17.0

Depends on D23365

Test Plan: {F7273745, size=full}

Reviewers: #plasma, broulik, meven, filipf

Reviewed By: meven

Subscribers: filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23368