Fix maximum y-value-range computation
Summary:
Rounding to the next 10 did not work if the maximum value was something like y_max = 30.947
In this case modulo 10 gives 0 and therefore the maximum values becomes 30 in the end, though it should be 40
otherwise the peak points outside of the chart.

Floor(max) + 10 always works.

Test Plan:
Before:
{F7369681}
After:
{F7369680}

Reviewers: #vdg, #plasma, ngraham

Reviewed By: #vdg, ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24029