refactor about-distro to more generically support textual dumps
Summary:
this previously relied on manually replicating strings and whatnot, it was
very awkward to read and the load functions were also getting a bit long.

introducing the new entry system:
generic table entries are now represented by a more generic Entry object
comprised of the entry label (e.g. 'Plasma Version:') and its value
(e.g. '5.17.0'). from those entries the UI is then constructed (i.e.
large parts of the UI are now no longer in the designer file but rather
constructed manually).
all visible entries are further more collected in a global list of entries
from which we'll generate the clipboard dump.
localization switching is now carried out through KLocalized String instead
of having string copies floating about. depending on context we'll either
resolve the localized or original string.

this should also make it easier to change the data set in the future, be
it adding or removing new data

RTL languages have a broken clipboard but it looks that was the case before
as well and I utterly failed to bend it into shape with bidirectional
control characters :(

Test Plan:
# LANGUAGE=en
- all entires shown
- copy to clipboard dumps all entries as before

# LANGUAGE=de
- as above
- copy to as english dumps in english rather than german

# LANGUAGE=he
- as above
- copy in hebrew seems incorrectly order (probably because bidi algorithm falls over from both the RTL label being in the same line as the LTR numbers)

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26667