don't hold initdness in a static var
with the move to systemsettings the module live times seem to have changed
so the statics survive a module switch and would prevent the module from
initializing again when viewed a second time.
instead simply look at m_entries to determine if we need to init. entries
are never empty after load, so it's a reliable indication for initdness.

BUG: 417466