make opengl module support multiple dri devices
Summary:
previously the code would only ever look at dri0 when that may not actually
be the device that is used. instead list all possible devices (I am
assuming in nvidia/intel combos the actual computations may happen on any
of them depending on configuration).
as a side effect we can no longer list a single kernel module but instead
list the module per-device.
the devices use cardN names on linux.
for the /proc based fallback we do not use cardN but simply N (to be
improved in master with localized string).

the legacy /proc/dri/ support was entirely removed. it's not a thing since
linux 3.12, which is from 2013.

on freebsd we continue to only support one device, in part because I have
no way to test multi-device support. the info is woefully underwhelming and
could as well not be there anyway TBH.

BUG: 417986
FIXED-IN: 5.18.5

Test Plan: vm with 3 cards show all of them with suitable info

Reviewers: apol

Reviewed By: apol

Subscribers: pino, zzag, apol, adridg, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27980