adjust smb kcm size to be sufficient vis a vis headers
Summary:
we use Stretch to get a somewhat uniform appearance between the two views
but Stretch is bit meh when it comes to actual sizing as it acts kind of
like QSizePolicy::Preferred in that it allows shrinking columns so small
that even the header gets cut off. this of course looks silly, so we
now set the largest required width as minimal section width. this prevents
the user from cutting off headers by resizing the window.

secondly, and entirely related to that, because KCMs are embedded into
a scrollview their sizehint kinda doesn't make any difference for
the window size which then combined with the aforementioned stretch problem
can lead to cut off headers by default as the window size is too small
making the stretch cut into header text.
to ensure the default size is reasonable let the view adjust to content
on first show, effectively making the window as large as needed
regardless of implicit hints.

BUG: 419786
FIXED-IN: 5.18.5

Test Plan:
- empty dialog
- dialog with only shares
- dialog with shares and mounts

Reviewers: ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28678