usb: rewrite device enumeration using libusb-1.0
Use the libusb-1.0 library to list the available devices, and get most
of the details for them. This replaces all the current platform-specific
reading code, mostly for Linux with bits for older BSDs. It is not clear
how much of the BSD code still works; anyway, libusb is ported to more
OSes than this KCM, so there is a net win.

The only platform specific code left is the reading of the manifacturer,
product, and serial number strings, and the number of channels; this is
done on Linux only looking at sysfs as done before, because getting them
via libusb requires opening the devices (which usually fails as regular
user).

This also fixes a memory leak: the old devices are properly deleted when
reading the new ones.