usb: improve usb.ids parsing
Tighten the regexps used, so they match the format of lines slightly
better (avoiding the need to check for matched lengths), and use
captures so we can get the parts instead of splitting the lines
manually.

Also, tighten the checks used to know which lines are being parsed,
making sure to parse only what is needed. This avoid language codes
being misparsed as subclasses.