usb: rework translations of USB classes & subclasses
The current translations for classes and subclasses come from a static
classes.i18n, which is supposed to be filled with strings from usb.ids.
However, this file has few issues:
- it has not been updated in a very long time, and half of its string
  do not exist anymore
- there is no indication of how it was created in the beginning, so
  keeping it up-to-date is a non-trivial job
- translations for classes and subclasses are mixed together

To improve the situation, introduce a script, i18n-from-usb-ids.py,
that can be used to update classes.i18n from a specified usb.ids file;
this way, it is much easier to keep these messages up-to-date regularly,
and the printed Date from usb.ids is an indication on the last time
classes.i18n was updated.

In addition, add different contexts for classes and subclasses: classes
follow the context previously introduced to translate the classes from
libusb.

As semi-related change: translate the class string only when reading it
from usb.ids, as we get translated strings from the mapping of the
libusb class.