repair the version header
this had likely been broken ever since the module got folded into
kinfocenter :(

it occurred to me that the previous code was daft anyway, that's why it
broke. the way we defined the variable would hide the cmake-side
variable being undefined/falsey. by using cmakedefine the macro is only
defined iff the variable is not undefined/falsey and as a consequence
the build would fail if the variable is undefined as then the
PROJECT_VERSION identifier is undeclared.

in conclusion the new code will fail if the variable is not defined,
explicitly undef, explicit OFF etc., ensuring that moving forward we
have a version.