unbreak about CPU value when solid is missing a product string
instead of `Processors: 4 x` one now gets `Processors: 4`, or
`Processors: 4 x Unknown Type, 12 x AMD Ryzen 5 3600X 6-Core Processor`
when some of the cores have a product and the others do not.

The latter is an incredible edge case but I suppose we best handle it
properly. It did require reshuffling of the entry though. We now list
the cores in the ctor to generate the label but only generate the values
at call time so as to localize this potential placeholder accordingly.

Meanwhile the former case ultimately lumps all processors together even
when they are actually different from a physical POV (such as the
heterogenous CPU set up of "ARM big.LITTLE"), but given the high level
overview this KCM provides there is no point in expansively trying to
figure out what the actual topology is (we do indeed not concern
ourselves with topology at all in this entry).

BUG: 439464
FIXED-IN: 5.22.4


(cherry picked from commit 0e5e4567141b814f67ef4e6971121ba9b922b265)