add a new advanced CPU module
renders lscpu, providing hugely more advanced info than we previously
had

this also serves as a jumping off point for turning the vulkan kcm into
generic tech to build CLI-based kcms.
unfortunately due to the nature c++kcms wire up with the kpackages to
supply the qml this does entail a fair amount of boilerplate. I'm not
sure there is a good way to get rid of this though.

the command runner class is now in a static lib that is pulled into all
modules and almost the entire KCM qml is shared as well. this leaves
only the metadata and some boilerplate code behind for each individual
kcm.

commandoutputcontext no longer implements qmlparserstatus because it no
longer gets constructed from qml making it always complete by default