port the output kcm from simplekcm to abstractkcm
this has a number of advantages:

- we no longer need to put our fields in the footer
- the QQC2 scollview actually has complete scrollbar coverage on all
sides
- the previous point then allows us to use NoWrap to prevent excessively
long lines to get wrapped when the format of the output doesn't really
need or want them this way (this is particularly useful for a
potentially upcoming interrupts kcm where we definitely don't want to
line break the format of /proc/interrupts to not make reading
unncessarily hard) -- or indeed the KCM can pick any wrap mode it finds
best for the given output
- textarea in combination with scrollview does automatically snap to a
selection area. this will ultimately make a proper search filed easier
to implement

to implement this our loaded state items now always fill the available
area and the actual display items are centered inside (e.g. the
placeholder message). this also addresses a review concern Nate had
because originally we were not able to place the placeholder properly