disable textarea background
this gives the kcms a "flat" look where the content of the area
visually becomes the content of the module. seeing as the module
has no other output or options this seems entirely reasonable and
offers a cleaner look over all.
this was originally brought up in the discussion of !46