delay busy indicator a tad
previously the indicator was shown as soon as the component is complete.
this is a bit meh because shortly after the output may have finished as
well and so the indicator would flash up for a couple draws and then
quickly disappear again.

instead only show the indicator if the loading state doesn't finish
within 500ms. this should be short enough a time to not have the empty
page be a concern