CommandOutputKCM: fix hacky background drawing
This hack is required because of a bug in SimpleKCM which is being fixed
in https://invent.kde.org/frameworks/kdeclarative/-/merge_requests/79.
With that fixed, we can simply use standard means of changing the
colors, such as overriding the colorSet. Ensuring that the footer has
the right color entails the use of a standard QQC2.ToolBar, which also
draws an appropriately-colored separator on the top when it is used as a
footer, because it was made for just this purpose.

There are no visual changes.