remove dma and ioport kcms + remainder of base infrastructure
a couple weeks ago we were briefly musing on this topic on Matrix but
couldn't really find a solid use case for these kcms KCMs as the chance
of a user wanting or needing to know dma or ioport info is quickly
approaching zero, even as far as diagnostics go they are very far out
there.

since they are the only thing holding up removal of the os base
infrastructure we'll drop the modules along with the remainder of the
base code.

if we found need for them again they should be easy enough to make from
scratch using the cli infrastructure instead. there's zero code reuse
potential for cli based modules either way.