De-duplicate metadata for QML KPackages
This way the embedded json metadata of the C++ plugins will be used instead.

By moving the main.qml file to a "ui"-subfolder we don't need to
configure the paths for the mainscript, but can just use the defaults.

To test this change, all the KCMs packages in "PREFIX/share/kpackage/kcms/"
should be removed. Otherwise we might have some leftovers lying around.