Move "Copy to Clipboard" action to a nice visible button
Right now it's possible to copy the text on a page that's based on
CommandOutputKCM, but this is hidden because the action only lives in a
context menu when you right-click on the view.

This commit fixes that by moving the action to a nice big visually
obvious button on the footer. This should make the action
discoverable--enough that we can remove the single-item context menu
that previously held the action.