about-distro: clarify scope of "Copy" Action
"Copy" actions generally rely on context or proximity to make clear
what will be copied. In this case we have none; the "Copy" button is
located far from the text that will be copied, and it lives on a toolbar
that's visually separated from the content.

Accordingly, let's clarify the scope of what it does by simply having it
tell you.