Avoid QML singletons to work with shared QML engines
KCM's were registering singleton instances. This has the lifespan of the
QMLEngine, and we were using objects with the lifespan of the loaded
KCM. Recent changes share the QML engine used between modules.

We have an existing mechanism to link the C++ container to the loaded UI
which is used in all other system settings. Port everything to using the
attached kcm object and adding a getter there.

BUG: 480881