CommandOutputContext: handle pkexec failure
If a command is run with pkexec and the
authentication fails or is cancelled, the
output will be the following message:

```
Error executing command as another user: Not authorized

This incident has been reported.
```

This is not a user-friendly message, and is not
currently treated as an error. Now we treat it as
an error and provide a more user-friendly message.