CommandOutputKCM: Use correct newlines for HTML-styled content
Right now it assumes plaintext and always uses a `\n`. But the Firmware
Security KCM displays HTML-styled text, where this won't work.

We need a way to tell the object that it's going to be displaying HTML
content, and use appropriate newlines. This commit implements that.

BUG: 499069
FIXED-IN: 6.3.0

Co-authored-by: Harald Sitter <sitter@kde.org>


(cherry picked from commit cd257602e2a4798b279846b131a355ad46001d97)

6665db3b CommandOutputKCM: Use correct newlines for HTML-styled content
6f77cbd8 Do it with enums
35d8f1b6 Try to make it work
16b8f7c0 Harald's better stuff

Co-authored-by: Nate Graham <nate@kde.org>