kcms/energy: Improve layout of the device selector
Port the device selector to an actual ListView instead of Row+Repeater
combo, so we can use the implicit height of the delegate, instead of
imposing a fixed height, which fails for bigger fonts.

Also improve the layout by centering the available buttons, and adapt
their width to the number of devices, so it doesn't need horizontal
scrolling and resembles more a Navigation bar

While at it, also simplify the code by using the ListView `currentItem`
to initialize and select the current device.

As an added benefit there is now keyboard navigation

BUG: 499290
BUG: 499350
FIXED-IN: 6.4