kcms/energy: Simplify the battery details model
To use required properties in a delegate, every item in the model
needs to provide the same full list of properties.

Due to the nature of the battery details, the details model was
setting different properties for each case, and custom conditional
handling to format the displayed text.

Let's simplify the model to have just a label, a source property
and a formatter function to handle how to display it.

As a benefit, we can reuse some formatter functions