about-distro: more aggressively fall back to opengl detection
this covers scenarios where the system may sport older hardware that
doesn't have vulkan support. these systems will report as llvmpipe on
the vulkan query because there is in fact no hardware support. when that
happens we should also not be getting a device type. in combination this
allows us to identify a scenario where fallback is called for